fixed https://github.com/typingArtist/linux_media/issues/1
Without this patch applied, saa716x_core takes all reference counts and leaves the
specific modules like saa716x_budget without a use tick. This leads to the situation
that while the application is running, saa716x_budget still has a reference count of
zero and can be unloaded, which freezes the kernel immediately. It is necessary for
dvb_register_adapter to get a reference to the real adapter-specific module to avoid
this case.